Update verification for comment controllers

+++ backend/controllers/commentController.js
@@ -1,4 +1,3 @@
-const validationResult = require('express-validator');
 const Comment = require('../models/comment');
 const verify = require('../config/passport');
 const asyncHandler = require('express-async-handler');
@@ -9,39 +8,59 @@ exports.comment_list = asyncHandler(async (req, res, next) => {
         .populate('user')
         .exec()
 
-    res.json(allComments);
+    if (!allComments) {
+        return res
+            .status(404)
+            .json({ error: true, msg: 'Comments unavailable' })
+    }
+
+    res.json({ error: false, allComments });
 });
 
 exports.comment_create_post = asyncHandler(async (req, res, next) => {
-    const errors = validationResult(req);
-    if (!errors.empty()) {
-        res.json(errors.array());
+    const { user, text, date_posted } = req.body;
+    if (!text) {
+        return res
+            .status(404)
+            .json({ error: true, msg: 'Text required' });
     }
 
     const createdComment = new Comment({
-        user: req.body.user,
-        text: req.body.text,
-        date_posted: Date.now()
+        user,
+        text,
+        date_posted
     });
 
     createdComment.save();
     verify(req.token, createdComment);
+    res.json({ error: false, createdComment });
 });
 
 exports.commentId_update = asyncHandler(async (req, res, next) => {
-    const errors = validationResult(req);
-    if (!errors.empty()) {
-        res.json(errors.array());
+    const comment = Comment.findOne(req.params.id);
+    if (!comment) {
+        return res
+            .status(404)
+            .json({ error: true, msg:'Comment not found' })
     }
-
-    const updatedComment = await Comment.findByIdAndUpdate(req.params.id);
     
+    const updatedComment = await Comment.findByIdAndUpdate(req.params.id);
+
     updatedComment.save();
     verify(req.token, updatedComment);
+    res.json({ error: false, updatedComment });
 });
 
 exports.commentId_delete = asyncHandler(async (req, res, next) => {
+    const comment = Comment.findOne(req.params.id);
+    if (!comment) {
+        return res
+            .status(404)
+            .json({ error: true, msg:'Comment not found' })
+    }
+
     const deletedComment = await Comment.findByIdAndDelete(req.params.id);
 
     verify(req.token, deletedComment);
+    res.json({ error: false, deletedComment });
 });
